+++ projects/igniteui-angular/src/lib/grid/grid.common.ts
@@ -1,4 +1,4 @@
-﻿import { DOCUMENT } from '@angular/common';
+﻿import { DOCUMENT, DatePipe, DecimalPipe } from '@angular/common';
 import {
     ChangeDetectorRef,
     Directive,
@@ -12,6 +12,8 @@ import {
     OnDestroy,
     OnInit,
     Output,
+    Pipe,
+    PipeTransform,
     Renderer2,
     TemplateRef
 } from '@angular/core';
@@ -598,6 +600,37 @@ export class IgxGroupAreaDropDirective extends IgxDropDirective {
             if (column.groupable && !isGrouped) {
                 column.grid.groupBy({ fieldName: column.field, dir: SortingDirection.Asc, ignoreCase: column.sortingIgnoreCase });
             }
+        }
+    }
+}
+
+/**
+ *@hidden
+ */
+@Pipe({
+    name: 'igxdate'
+})
+export class IgxDatePipeComponent extends DatePipe implements PipeTransform {
+    transform(value: any): string {
+        if (value && value instanceof Date) {
+            return super.transform(value);
+        } else {
+            return value;
+        }
+    }
+}
+/**
+ *@hidden
+ */
+@Pipe({
+    name: 'igxdecimal'
+})
+export class IgxDecimalPipeComponent extends DecimalPipe implements PipeTransform {
+    transform(value: any): string {
+        if (value && typeof value === 'number') {
+            return super.transform(value);
+        } else {
+            return value;
         }
     }
 }
