+++ projects/igniteui-angular/src/lib/grid/grid.component.ts
@@ -1,4 +1,4 @@
-import { DOCUMENT, DatePipe, DecimalPipe } from '@angular/common';
+import { DOCUMENT } from '@angular/common';
 import {
     AfterContentInit,
     AfterViewInit,
@@ -19,8 +19,6 @@ import {
     NgZone,
     OnDestroy,
     OnInit,
-    Pipe,
-    PipeTransform,
     Output,
     QueryList,
     TemplateRef,
@@ -126,38 +124,6 @@ export interface IColumnMovingEndEventArgs {
     source: IgxColumnComponent;
     target: IgxColumnComponent;
     cancel: boolean;
-}
-
-/**
- *@hidden
- */
-@Pipe({
-    name: 'igxdate'
-})
-export class IgxDatePipeComponent extends DatePipe implements PipeTransform {
-    transform(value: any): string {
-        if (value && value instanceof Date) {
-            return super.transform(value);
-        } else {
-            return value;
-        }
-    }
-}
-
-/**
- *@hidden
- */
-@Pipe({
-    name: 'igxdecimal'
-})
-export class IgxDecimalPipeComponent extends DecimalPipe implements PipeTransform {
-    transform(value: any): string {
-        if (value && typeof value === 'number') {
-            return super.transform(value);
-        } else {
-            return value;
-        }
-    }
 }
 
 /**
